+++ src/test/java/integration/reservationtime/ReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package integration.reservation;
+package integration.reservationtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
